pyproject.toml
[project]
name = "fragberta"
version = "0.1.0"
requires-python = "==3.10.*"
dependencies = [
    # Core ML
    "transformers==4.57.1",
    "datasets==4.4.1",
    "accelerate==1.11.0",
    "tokenizers==0.22.1",
    "safetensors==0.6.2",
    "huggingface-hub==0.36.0",
    # PyTorch — version managed via tool.uv.sources below
    "torch==2.5.1",
    "torchvision==0.20.1",
    # Chemistry
    "rdkit==2025.9.1",
    "chemprop==1.6.1",
    "datamol==0.12.5",
    "safe-mol==0.1.14",
    "selfies==2.2.0",
    # HPO & tracking
    "optuna==4.6.0",
    "hyperopt==0.2.7",
    "wandb==0.23.0",
    "tensorboardx==2.6.4",
    # Numerics
    "numpy>=1.24",
    "scipy==1.10.1",
    "scikit-learn==1.7.2",
    "pandas==2.3.3",
    "pyarrow==22.0.0",
    # Jupyter
    "jupyterlab==4.4.10",
    "ipython==8.37.0",
    "ipykernel==7.1.0",
    # Viz
    "matplotlib==3.10.7",
    "seaborn==0.13.2",
    # Utils
    "tqdm==4.67.1",
    "pyyaml==6.0.3",
    "loguru==0.7.3",
    "rich==14.2.0",
    "pydantic==2.12.4",
    "click==8.3.0",
    "typer==0.20.0",
    "requests==2.32.5",
    "httpx==0.28.1",
    "aiohttp==3.13.2",
    "flask==3.1.2",
    "networkx==3.4.2",
    "dask==2025.11.0",
    "xarray==2025.6.1",
    "sqlalchemy==2.0.45",
    "alembic==1.17.2",
    "pillow==12.0.0",
    "psutil==7.1.3",
    "gitpython==3.1.45",
    "evaluate==0.4.6",
    "triton==3.1.0",
    "sympy==1.13.1",
    "beautifulsoup4==4.14.2",
    # Sphinx
    "sphinx==8.1.3",
    # ... other packages as needed
    "setuptools>=82.0.0",
    "pkgutil-resolve-name>=1.3.10",
]

[tool.uv.sources]
# Change cu121 → cu118 or cu130 depending on the cluster
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.ruff]
line-length = 88  # match black

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle, pyflakes, isort
ignore = ["E501"]